requirements.txt
arrow==1.2.3
asgiref==3.7.2
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
item==0.0.3
lxml==4.9.3
numpy==1.25.1
pandas==2.0.3
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytils==0.4.1
pytz==2023.3
redis==3.5.3
requests==2.31.0
selection==0.0.21
six==1.16.0
soupsieve==2.4.1
sourcery==1.6.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
user-agent==0.1.10
wcwidth==0.2.6
weblib==0.1.30
